+++ CHANGELOG.md
@@ -9,6 +9,7 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 ## [Unreleased]
 
+- Add excluded resources environment variable.
 - Added create/cancel bookings endpoints and tests.
 - Updated GitHub workflow images.